Fin projet

+++ src/main/java/com/ninja_squad/geektic/domaine/CentreInteret.java
@@ -12,17 +12,17 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 public class CentreInteret {
 	@Id
 	@Column(name = "idcentreinteret")
-	private Long id;
+	private int id;
 	private String libelle;
 	
 	@JsonIgnore
 	@ManyToMany(mappedBy = "centreinterets") 
 	private List<Utilisateur> utilisateurs;
 	
-	public Long getId() {
+	public int getId() {
 		return id;
 	}
-	public void setId(Long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
 	public String getLibelle() {
@@ -31,5 +31,4 @@ public class CentreInteret {
 	public void setLibelle(String libelle) {
 		this.libelle = libelle;
 	}
-
 }
